app/requirements.txt
fastapi==0.70.0
python-socketio==5.5.0
uvicorn==0.15.0
python-dotenv==0.19.2
pydantic==1.9.0
aiohttp==3.8.1 